Refactor: Move UI components into src/ui package for better maintainability
