+++ .gitignore
@@ -12,3 +12,6 @@ computed_data/
 
 .DS_Store
 .idea
+
+CLAUDE.md
+.claude/
