+++ src/ui/base.py
@@ -1,0 +1,4 @@
+class BaseComponent:
+    def on_resize(self, window): pass
+    def draw(self, window): pass
+    def on_mouse_press(self, window, x: float, y: float, button: int, modifiers: int) -> bool: return False
