+++ src/ui/lap_times.py
@@ -1,0 +1,88 @@
+import arcade
+from typing import List
+from src.lib.time import format_time
+from .base import BaseComponent
+
+class LapTimeLeaderboardComponent(BaseComponent):
+    def __init__(self, x: int, right_margin: int = 260, width: int = 240):
+        self.x = x
+        self.width = width
+        self.entries = []  # list of dicts: {'pos', 'code', 'color', 'time'}
+        self.rects = []    # clickable rects per entry
+        self.selected = []  # Changed to list
+        self.row_height = 25
+        self._visible = True
+
+    def set_entries(self, entries: List[dict]):
+        """Accept a list of dicts with keys: pos, code, color, time"""
+        self.entries = entries or []
+    
+    @property
+    def visible(self) -> bool:
+        return self._visible
+    
+    @visible.setter
+    def visible(self, value: bool):
+        self._visible = value
+    
+    def toggle_visibility(self) -> bool:
+        """
+        Toggle the visibility of the progress bar
+        """
+        self._visible = not self._visible
+        return self._visible
+
+    def draw(self, window):
+        # Skip rendering entirely if hidden
+        if not self._visible:
+            return
+        self.selected = getattr(window, "selected_drivers", [])
+        leaderboard_y = window.height - 40
+        arcade.Text("Lap Times", self.x, leaderboard_y, arcade.color.WHITE, 20, bold=True, anchor_x="left", anchor_y="top").draw()
+        self.rects = []
+        for i, entry in enumerate(self.entries):
+            pos = entry.get('pos', i + 1)
+            code = entry.get('code', '')
+            color = entry.get('color', arcade.color.WHITE)
+            time_str = entry.get('time', '')
+            current_pos = i + 1
+            top_y = leaderboard_y - 30 - ((current_pos - 1) * self.row_height)
+            bottom_y = top_y - self.row_height
+            left_x = self.x
+            right_x = self.x + self.width
+            # store clickable rect (code, left, bottom, right, top)
+            self.rects.append((code, left_x, bottom_y, right_x, top_y))
+
+            # selection highlight
+            if code in self.selected:
+                rect = arcade.XYWH((left_x + right_x) / 2, (top_y + bottom_y) / 2, right_x - left_x, top_y - bottom_y)
+                arcade.draw_rect_filled(rect, arcade.color.LIGHT_GRAY)
+                text_color = arcade.color.BLACK
+            else:
+                # accept tuple rgb or fallback to white
+                text_color = tuple(color) if isinstance(color, (list, tuple)) else arcade.color.WHITE
+
+            # Draw code on left, time right-aligned
+            arcade.Text(f"{pos}. {code}", left_x + 8, top_y, text_color, 16, anchor_x="left", anchor_y="top").draw()
+            arcade.Text(time_str, right_x - 8, top_y, text_color, 14, anchor_x="right", anchor_y="top").draw()
+
+    def on_mouse_press(self, window, x: float, y: float, button: int, modifiers: int):
+        for code, left, bottom, right, top in self.rects:
+            if left <= x <= right and bottom <= y <= top:
+                is_multi = (modifiers & arcade.key.MOD_SHIFT)
+
+                if is_multi:
+                    if code in self.selected:
+                        self.selected.remove(code)
+                    else:
+                        self.selected.append(code)
+                else:
+                    if len(self.selected) == 1 and self.selected[0] == code:
+                        self.selected = []
+                    else:
+                        self.selected = [code]
+
+                window.selected_drivers = self.selected
+                window.selected_driver = self.selected[-1] if self.selected else None
+                return True
+        return False
